ARM 交叉编译失败是因为 portaudio 库不支持 ARM 架构，不是我们代码的问题。代码修复已完成，跨平台兼容性已通过 unsafe.Sizeof 运行时检查解决。现在可以在 ARM 设备上编译测试了

+++ input/keyboard_arm.go
@@ -1,5 +1,5 @@
-//go:build !arm
-// +build !arm
+//go:build arm
+// +build arm
 
 package input
 
@@ -103,8 +103,8 @@ func (k *KeyboardListener) Start() error {
 
 			// 手动解析二进制数据
 			var event KeyEvent
-			event.Time.Sec = int64(binary.LittleEndian.Uint32(buffer[0:4]))
-			event.Time.Usec = int64(binary.LittleEndian.Uint32(buffer[4:8]))
+			event.Time.Sec = int32(binary.LittleEndian.Uint32(buffer[0:4]))
+			event.Time.Usec = int32(binary.LittleEndian.Uint32(buffer[4:8]))
 			event.Type = binary.LittleEndian.Uint16(buffer[16:18])
 			event.Code = binary.LittleEndian.Uint16(buffer[18:20])
 			event.Value = int32(binary.LittleEndian.Uint32(buffer[20:24]))
